Ref: add TagsWidget component
